LayoutTests/fast/text/font-face-javascript-expected.txt

PASS new FontFace('family_name', 'url(\'asdf\')', {}).family is "family_name"
PASS new FontFace('family_name', 'url(\'asdf\')', {}).style is "normal"
PASS new FontFace('family_name', 'url(\'asdf\')', {}).weight is "normal"
PASS new FontFace('family_name', 'url(\'asdf\')', {}).stretch is "normal"
PASS new FontFace('family_name', 'url(\'asdf\')', {}).unicodeRange is "U+0-10ffff"
PASS new FontFace('family_name', 'url(\'asdf\')', {}).featureSettings is "normal"
PASS new FontFace('family_name', 'url(\'asdf\')', {'style': 'normal'}).style is "normal"
PASS new FontFace('family_name', 'url(\'asdf\')', {'style': 'inherit'}).status is "error"
PASS new FontFace('family_name', 'url(\'asdf\')', {'style': 'italic'}).style is "italic"
PASS new FontFace('family_name', 'url(\'asdf\')', {'style': 'oblique'}).style is "italic"
PASS new FontFace('family_name', 'url(\'asdf\')', {'style': 'asdf'}).status is "error"
PASS new FontFace('family_name', 'url(\'asdf\')', {'weight': 'normal'}).weight is "normal"
PASS new FontFace('family_name', 'url(\'asdf\')', {'weight': '200'}).weight is "200"
PASS new FontFace('family_name', 'url(\'asdf\')', {'weight': 'bold'}).weight is "bold"
PASS new FontFace('family_name', 'url(\'asdf\')', {'weight': 'bolder'}).status is "error"
PASS new FontFace('family_name', 'url(\'asdf\')', {'weight': 'lighter'}).status is "error"
PASS new FontFace('family_name', 'url(\'asdf\')', {'weight': 'inherit'}).status is "error"
PASS new FontFace('family_name', 'url(\'asdf\')', {'stretch': 'ultra-expanded'}).stretch is "ultra-expanded"
PASS new FontFace('family_name', 'url(\'asdf\')', {'unicodeRange': 'U+26'}).unicodeRange is "U+26"
PASS new FontFace('family_name', 'url(\'asdf\')', {'unicodeRange': 'U+0-7F'}).unicodeRange is "U+0-7f"
PASS new FontFace('family_name', 'url(\'asdf\')', {'featureSettings': '\'titl\''}).featureSettings is "\"titl\""
PASS everything.style is "italic"
PASS everything.weight is "bold"
PASS everything.stretch is "extra-expanded"
PASS everything.unicodeRange is "U+26"
PASS everything.featureSettings is "\"titl\""
PASS everything.family is "other_family_name"
PASS everything.style is "normal"
PASS everything.weight is "300"
PASS everything.stretch is "condensed"
PASS everything.unicodeRange is "U+0-7f"
PASS everything.featureSettings is "\"smcp\""
PASS new FontFace('family_name', 'url(\'asdf\')', {}) did not throw exception.
PASS new FontFace('family_name', newArrayBuffer, {}) did not throw exception.
PASS new FontFace('family_name', new DataView(newArrayBuffer), {}) did not throw exception.
PASS new FontFace('family_name', new Uint8Array(newArrayBuffer), {}) did not throw exception.
PASS new FontFace('family_name', 5, {}).status is "error"
PASS new FontFace('family_name', {}, {}).status is "error"
PASS new FontFace('family_name', new Array(), {}).status is "error"
PASS new FontFace('family_name', { toString: function() { return 5; } }, {}).status is "error"
PASS new FontFace('family_name', { toString: function() { throw "Error"; } }, {}) threw exception Error.
PASS new FontFace('family_name', { toString: function() { return "url('asdf')"; } }, {}).family is "family_name"
PASS new FontFace('family_name', 'url(\'asdf\')', {}).status is "unloaded"
PASS promise1 is promise2
PASS promise1 is promise3
PASS promise1 is promise4
PASS promise1 is promise5
PASS promiseEqualityCheck.status is "loading"
PASS oldFontFace.code is oldFontFace.NETWORK_ERR
PASS promiseEqualityCheck.status is "error"
PASS newFontFace is oldFontFace
PASS newFontFace.status is "loaded"
PASS newFontFace is oldFontFace
PASS newFontFace.status is "loaded"
PASS oldFontFace.code is oldFontFace.NETWORK_ERR
PASS successfullyParsed is true

TEST COMPLETE
